Updated basic grab/drop.

+++ src/controllers/base.ts
@@ -1,8 +1,6 @@
 import {
     AbstractMesh,
-    PhysicsMotionType,
     Scene,
-    TransformNode,
     Vector3,
     WebXRControllerComponent,
     WebXRDefaultExperience,
@@ -12,24 +10,19 @@ import {DiagramManager} from "../diagram/diagramManager";
 import {DiagramEvent, DiagramEventType} from "../diagram/types/diagramEntity";
 import log from "loglevel";
 import {ControllerEventType, Controllers} from "./controllers";
-import {toDiagramEntity} from "../diagram/functions/toDiagramEntity";
-import {setupTransformNode} from "./functions/setupTransformNode";
-import {reparent} from "./functions/reparent";
-import {snapGridVal} from "../util/functions/snapGridVal";
-import {snapRotateVal} from "../util/functions/snapRotateVal";
 import {grabAndClone} from "./functions/grabAndClone";
-import {isDiagramEntity} from "../diagram/functions/isDiagramEntity";
 import {ClickMenu} from "../menus/clickMenu";
 import {displayDebug} from "../util/displayDebug";
-import {beforeRenderObserver} from "./functions/beforeRenderObserver";
 import {motionControllerObserver} from "./functions/motionControllerObserver";
-import {handleWasGrabbed} from "./functions/handleWasGrabbed";
-import {buildDrop} from "./functions/buildDrop";
-import {pointable} from "./functions/pointable";
 import {DefaultScene} from "../defaultScene";
 import {DiagramEventObserverMask} from "../diagram/types/diagramEventObserverMask";
+import {DiagramObject} from "../objects/diagramObject";
+import {snapAll} from "./functions/snapAll";
+import {MeshTypeEnum} from "../diagram/types/meshTypeEnum";
+import {getMeshType} from "./functions/getMeshType";
 
 const CLICK_TIME = 300;
+
 
 
 export class Base {
@@ -37,20 +30,18 @@ export class Base {
     protected xrInputSource: WebXRInputSource;
     protected speedFactor = 4;
     protected readonly scene: Scene;
+    protected grabbedObject: DiagramObject = null;
     protected grabbedMesh: AbstractMesh = null;
-    protected grabbedMeshParentId: string = null;
-    protected previousParentId: string = null;
-    protected previousRotation: Vector3 = null;
-    protected previousScaling: Vector3 = null;
-    protected previousPosition: Vector3 = null;
+    protected grabbedMeshType: MeshTypeEnum = null;
     private clickStart: number = 0;
     protected readonly xr: WebXRDefaultExperience;
     protected readonly diagramManager: DiagramManager;
-    private lastPosition: Vector3 = null;
     protected controllers: Controllers;
     private clickMenu: ClickMenu;
     private pickPoint: Vector3 = new Vector3();
+    private meshUnderPointer: AbstractMesh;
     private logger = log.getLogger('Base');
+
     constructor(controller: WebXRInputSource,
                 xr: WebXRDefaultExperience,
                 diagramManager: DiagramManager) {
@@ -60,15 +51,14 @@ export class Base {
         this.scene = DefaultScene.Scene;
         this.xr = xr;
         this.scene.onPointerObservable.add((pointerInfo) => {
-            if (pointerInfo.pickInfo.pickedMesh?.metadata?.template) {
-                const mesh = pointerInfo.pickInfo.pickedMesh;
-                //const pos = mesh.absolutePosition;
-                this.pickPoint.copyFrom(pointerInfo.pickInfo.pickedPoint);
+            if (pointerInfo.pickInfo.pickedMesh) {
+                this.meshUnderPointer = pointerInfo.pickInfo.pickedMesh;
+                if (this.diagramManager.getDiagramObject(pointerInfo.pickInfo.pickedMesh.id)) {
+                    this.pickPoint.copyFrom(pointerInfo.pickInfo.pickedPoint);
+                }
             }
-
         });
         this.diagramManager = diagramManager;
-        this.scene.onBeforeRenderObservable.add(beforeRenderObserver, -1, false, this);
 
         //@TODO THis works, but it uses initGrip, not sure if this is the best idea
         this.xrInputSource.onMotionControllerInitObservable.add(motionControllerObserver, -1, false, this);
@@ -77,7 +67,7 @@ export class Base {
             switch (event.type) {
                 case ControllerEventType.PULSE:
                     if (event.gripId == this?.xrInputSource?.grip?.id) {
-                        this.xrInputSource?.motionController?.pulse(.25, 30)
+                        this.xrInputSource?.motionController?.pulse(.35, 50)
                             .then(() => {
                                 this.logger.debug("pulse done");
                             });
@@ -115,7 +105,9 @@ export class Base {
                         window.setTimeout(() => {
                             if (this.clickStart > 0) {
                                 this.logger.debug("grabbing and cloning");
-                                this.grab(true);
+                                const clone = grabAndClone(this.diagramManager, this.meshUnderPointer, this.xrInputSource.motionController.rootMesh);
+                                this.grabbedObject = clone;
+                                this.grabbedMesh = clone.mesh;
                             }
                         }, 300, this);
                     }
@@ -123,67 +115,41 @@ export class Base {
                     const clickEnd = Date.now();
                     if (this.clickStart > 0 && (clickEnd - this.clickStart) < CLICK_TIME) {
                         this.click();
+                    } else {
+                        if (this.grabbedObject || this.grabbedMesh) {
+                            this.drop();
+                        }
                     }
-                    this.drop();
                     this.clickStart = 0;
                 }
             }
         }, -1, false, this);
     }
 
-    private grab(cloneEntity: boolean = false) {
-        let mesh = this.xr.pointerSelection.getMeshUnderPointer(this.xrInputSource.uniqueId);
 
+    private grab() {
+        let mesh = this.xr.pointerSelection.getMeshUnderPointer(this.xrInputSource.uniqueId);
         if (!mesh) {
             return;
         }
-        let player = false;
+        this.grabbedMesh = mesh;
+        this.grabbedMeshType = getMeshType(mesh, this.diagramManager);
+
         displayDebug(mesh);
-        if (!isDiagramEntity(mesh)) {
-            if (handleWasGrabbed(mesh)) {
-                mesh && mesh.setParent(this.xrInputSource.motionController.rootMesh);
-                this.grabbedMesh = mesh;
-            } else {
-                if (mesh?.metadata?.grabbable) {
-                    this.grabbedMesh = mesh;
-                } else {
-                    return;
-                }
-            }
-        } else {
-            if (mesh?.metadata?.template == '#connection-template') {
-                return;
-            }
-        }
+        switch (this.grabbedMeshType) {
+            case MeshTypeEnum.ENTITY:
+                const diagramObject = this.diagramManager.getDiagramObject(mesh.id);
+                diagramObject.baseTransform.setParent(this.xrInputSource.motionController.rootMesh);
+                this.grabbedObject = diagramObject;
+                break;
+            case MeshTypeEnum.HANDLE:
+                this.grabbedMesh.setParent(this.xrInputSource.motionController.rootMesh);
+                break;
+            case MeshTypeEnum.TOOL:
+                const clone = grabAndClone(this.diagramManager, mesh, this.xrInputSource.motionController.rootMesh);
+                this.grabbedObject = clone;
+                this.grabbedMesh = clone.mesh;
 
-        this.previousParentId = mesh?.parent?.id;
-        this.logger.warn("grabbed " + mesh?.id + " parent " + this.previousParentId);
-        this.previousRotation = mesh?.rotation.clone();
-        this.previousScaling = mesh?.scaling.clone();
-        this.previousPosition = mesh?.position.clone();
-
-        if ((!mesh.metadata?.grabClone || player) && !cloneEntity) {
-            if (mesh.physicsBody) {
-                const transformNode = setupTransformNode(mesh, this.xrInputSource.motionController.rootMesh);
-                mesh.physicsBody.setMotionType(PhysicsMotionType.ANIMATED);
-                this.grabbedMeshParentId = transformNode.id;
-            } else {
-                mesh.setParent(this.xrInputSource.motionController.rootMesh);
-            }
-            this.grabbedMesh = mesh;
-        } else {
-            this.logger.debug("cloning " + mesh?.id);
-            const clone = grabAndClone(this.diagramManager, mesh, this.xrInputSource.motionController.rootMesh);
-            clone.newMesh.metadata.grabClone = false;
-            clone.newMesh.metadata.tool = false;
-            this.grabbedMeshParentId = clone.transformNode.id;
-            this.grabbedMesh = clone.newMesh;
-            this.previousParentId = null;
-            const event: DiagramEvent = {
-                type: DiagramEventType.ADD,
-                entity: toDiagramEntity(clone.newMesh)
-            }
-            this.diagramManager.onDiagramEventObservable.notifyObservers(event, DiagramEventObserverMask.ALL);
         }
     }
 
@@ -192,58 +158,58 @@ export class Base {
         if (!mesh) {
             return;
         }
-        if (handleWasGrabbed(mesh)) {
-            mesh.setParent(this.scene.getMeshByName("platform"));
-            const location = {
-                position: {x: mesh.position.x, y: mesh.position.y, z: mesh.position.z},
-                rotation: {x: mesh.rotation.x, y: mesh.rotation.y, z: mesh.rotation.z}
-            }
-            localStorage.setItem(mesh.id, JSON.stringify(location));
-            return;
-        }
-        reparent(mesh, this.previousParentId, this.grabbedMeshParentId);
-        this.grabbedMeshParentId = null;
-        if (!mesh.physicsBody) {
-            const transform = new TransformNode('temp', this.scene);
-            transform.position = this.pickPoint;
-            mesh.setParent(transform);
-            mesh.rotation = snapRotateVal(mesh.rotation, this.diagramManager._config.current.rotateSnap);
-            transform.position = snapGridVal(transform.position, this.diagramManager._config.current.gridSnap);
-            mesh.setParent(null);
-            mesh.position = snapGridVal(mesh.position, this.diagramManager._config.current.gridSnap);
-            //mesh.position = snapGridVal(mesh.position, this.diagramManager._config.current.gridSnap);
-            //mesh.setPivotPoint(transform.position, Space.WORLD)
+        const diagramObject = this.grabbedObject;
+        switch (this.grabbedMeshType) {
+            case MeshTypeEnum.ENTITY:
+                if (diagramObject) {
+                    diagramObject.baseTransform.setParent(null);
+                    snapAll(this.grabbedObject.baseTransform, this.diagramManager.config, this.pickPoint);
+                    diagramObject.mesh.computeWorldMatrix(true);
+                    const event: DiagramEvent =
+                        {
+                            type: DiagramEventType.DROP,
+                            entity: diagramObject.diagramEntity
+                        }
+                    this.diagramManager.onDiagramEventObservable.notifyObservers(event, DiagramEventObserverMask.ALL);
+                    diagramObject.mesh.computeWorldMatrix(false);
+                }
 
-
-            //transform.dispose();
+                this.grabbedObject = null;
+                this.grabbedMesh = null;
+                this.grabbedMeshType = null;
+                break;
+            case MeshTypeEnum.TOOL:
+                this.grabbedObject.baseTransform.setParent(null);
+                snapAll(this.grabbedObject.baseTransform, this.diagramManager.config, this.pickPoint);
+                diagramObject.mesh.computeWorldMatrix(true);
+                const event: DiagramEvent =
+                    {
+                        type: DiagramEventType.DROP,
+                        entity: diagramObject.diagramEntity
+                    }
+                this.diagramManager.onDiagramEventObservable.notifyObservers(event, DiagramEventObserverMask.ALL);
+                diagramObject.mesh.computeWorldMatrix(false);
+                this.grabbedObject = null;
+                this.grabbedMesh = null;
+                this.grabbedMeshType = null;
+                break;
+            case MeshTypeEnum.HANDLE:
+                mesh.setParent(this.scene.getMeshByName("platform"));
+                const location = {
+                    position: {x: mesh.position.x, y: mesh.position.y, z: mesh.position.z},
+                    rotation: {x: mesh.rotation.x, y: mesh.rotation.y, z: mesh.rotation.z}
+                }
+                localStorage.setItem(mesh.id, JSON.stringify(location));
+                this.grabbedMesh = null;
+                this.grabbedMeshType = null;
+                this.grabbedObject = null;
+                break;
         }
-        this.previousParentId = null;
-        this.previousScaling = null;
-        this.previousRotation = null;
-        this.previousPosition = null;
-        this.grabbedMesh = null;
-        if (isDiagramEntity(mesh) && (mesh?.metadata?.template.indexOf('#') == -1)) {
-            return;
-        }
-        const event: DiagramEvent = buildDrop(mesh);
-
-        const body = mesh?.physicsBody;
-        if (body) {
-            body.setMotionType(PhysicsMotionType.DYNAMIC);
-            this.logger.debug(body.transformNode.absolutePosition);
-            this.logger.debug(this.lastPosition);
-            if (this.lastPosition) {
-                body.setLinearVelocity(body.transformNode.absolutePosition.subtract(this.lastPosition).scale(20));
-                //body.setLinearVelocity(this.lastPosition.subtract(body.transformNode.absolutePosition).scale(20));
-                this.logger.debug(this.lastPosition.subtract(body.transformNode.absolutePosition).scale(20));
-            }
-        }
-        this.diagramManager.onDiagramEventObservable.notifyObservers(event, DiagramEventObserverMask.ALL);
     }
 
     private click() {
         let mesh = this.xr.pointerSelection.getMeshUnderPointer(this.xrInputSource.uniqueId);
-        if (pointable(mesh)) {
+        if (this.diagramManager.isDiagramObject(mesh)) {
             this.logger.debug("click on " + mesh.id);
             if (this.clickMenu && !this.clickMenu.isDisposed) {
                 if (this.clickMenu.isConnecting) {
@@ -253,12 +219,9 @@ export class Base {
             } else {
                 this.clickMenu = this.diagramManager.diagramMenuManager.createClickMenu(mesh, this.xrInputSource);
             }
-
         } else {
             this.logger.debug("click on nothing");
         }
-
-
     }
 
     private initGrip(grip: WebXRControllerComponent) {
